Move most of the importing code out of import.py
